feat: add customer support chatbot. fix: add missing meta descriptions && correct canny page issues. ci: sitemap configuration
