+++ next-sitemap.config.js
@@ -2,6 +2,9 @@ module.exports = {
   siteUrl: process.env.NEXT_PUBLIC_SITE_ROOT,
   generateRobotsTxt: true,
   robotsTxtOptions: {
-      policies: [{ userAgent: '*', allow: '/' }]
+    policies: [{ userAgent: '*', allow: '/' }],
+    additionalSitemaps: [
+      process.env.NEXT_PUBLIC_DOCUMENTATION_URL + '/sitemap.xml', // <==== Add here
+    ],
   },
 };
