+++ src/app/about/page.tsx
@@ -8,7 +8,6 @@ export const metadata: Metadata = {
   title: 'About Us - Service as a Callgent',
   description: 'Discover how Callgent is revolutionizing the way users interact with systems through innovative solutions. Learn about our mission, team, and the technology that drives us.',
   keywords: 'Callgent, about Callgent, user-system interaction, technology, innovation, team',
-  // other metadata
 };
 
 const AboutPage = () => {
