+++ src/app/contact/page.tsx
@@ -7,7 +7,6 @@ export const metadata: Metadata = {
   title: 'Connect with Our Team - Service as a Callgent',
   description: 'Reach out to the Callgent team for inquiries, support, or feedback. Find our email, phone number, and fill out the contact form to get in touch with us directly.',
   keywords: 'Callgent, contact, support, feedback, get in touch, customer service',
-  // other metadata
 };
 
 const ContactPage = () => {
@@ -17,7 +16,6 @@ const ContactPage = () => {
         pageName="Contact Page"
         description="Lorem ipsum dolor sit amet, consectetur adipiscing elit. In varius eros eget sapien consectetur ultrices. Ut quis dapibus libero."
       />
-
       <Contact />
     </>
   );
