+++ src/app/layout.tsx
@@ -16,6 +16,7 @@ import Head from 'next/head';
 import { SpeedInsights } from '@vercel/speed-insights/next';
 import { Analytics } from "@vercel/analytics/react";
 import { GoogleTagManager } from '@next/third-parties/google';
+import Script from 'next/script';
 const Setting = ({ children }) => {
   useEffect(() => {
     const lang = getCookie('lang');
@@ -61,6 +62,20 @@ export default function RootLayout({ children }: Readonly<{
           </Setting>
         </Providers>
       </body>
+      <Script
+        id='tawk'
+        strategy="afterInteractive"
+        dangerouslySetInnerHTML={{
+          __html: `var Tawk_API=Tawk_API||{}, Tawk_LoadStart=new Date();(function(){var s1=document.createElement("script"),s0=document.getElementsByTagName("script")[0];s1.async=true;s1.src='https://embed.tawk.to/6641de0a9a809f19fb303daa/1htokmmh3';s1.charset='UTF-8';s1.setAttribute('crossorigin','*');s0.parentNode.insertBefore(s1,s0);})();`,
+        }}
+      />
+      <Script
+        id='statistics'
+        strategy="afterInteractive"
+        dangerouslySetInnerHTML={{
+          __html: `(function(c,l,a,r,i,t,y){ c[a]=c[a]||function(){(c[a].q=c[a].q||[]).push(arguments)}; t=l.createElement(r);t.async=1;t.src="https://www.clarity.ms/tag/"+i; y=l.getElementsByTagName(r)[0];y.parentNode.insertBefore(t,y); })(window, document, "clarity", "script", "m9rdd2ip30");`,
+        }}
+      />
     </html>
   );
 }
