+++ src/app/providers.tsx
@@ -1,4 +1,3 @@
-'use client';
 import React from 'react';
 import { ThemeProvider } from 'next-themes';
 import { Provider } from 'react-redux';
