+++ src/components/Common/Breadcrumb.tsx
@@ -8,8 +8,7 @@ const Breadcrumb = ({
   description: string;
 }) => {
   return (
-    <>
-      <section className="relative z-10 overflow-hidden pt-28 lg:pt-[150px]">
+    <section className="relative z-10 overflow-hidden pt-28 lg:pt-[150px]">
         <div className="container">
           <div className="-mx-4 flex flex-wrap items-center">
             <div className="w-full px-4 md:w-8/12 lg:w-7/12">
@@ -118,7 +117,6 @@ const Breadcrumb = ({
           </span>
         </div>
       </section>
-    </>
   );
 };
